+++ test.js
@@ -5,20 +5,18 @@ var timestring = require('./index');
 
 describe('timestring', function() {
   it('can parse a timestring', function() {
-    var ts = new timestring();
-
-    expect(ts.parse('1s')).to.equal(1);
-    expect(ts.parse('1m')).to.equal(60);
-    expect(ts.parse('1h')).to.equal(3600);
-    expect(ts.parse('1d')).to.equal(86400);
-    expect(ts.parse('1w')).to.equal(604800);
-    expect(ts.parse('1mth')).to.equal(2419200);
-    expect(ts.parse('1y')).to.equal(29030400);
+    expect(timestring('1s')).to.equal(1);
+    expect(timestring('1m')).to.equal(60);
+    expect(timestring('1h')).to.equal(3600);
+    expect(timestring('1d')).to.equal(86400);
+    expect(timestring('1w')).to.equal(604800);
+    expect(timestring('1mth')).to.equal(2419200);
+    expect(timestring('1y')).to.equal(29030400);
   });
 
   it('can return a value in a specified unit', function() {
-    expect((new timestring()).parse('5m', 's')).to.equal(300);
-    expect((new timestring()).parse('5m', 'm')).to.equal(5);
+    expect(timestring('5m', 's')).to.equal(300);
+    expect(timestring('5m', 'm')).to.equal(5);
   });
 
   it('uses the passed options instead of the defaults', function() {
@@ -29,35 +27,17 @@ describe('timestring', function() {
       monthsPerYear: 4
     };
 
-    var ts = new timestring(opts);
-
-    expect(ts.parse('1d', 'h')).to.equal(1);
-    expect(ts.parse('1w', 'd')).to.equal(2);
-    expect(ts.parse('1mth', 'w')).to.equal(3);
-    expect(ts.parse('1y', 'mth')).to.equal(4);
+    expect(timestring('1d', 'h', opts)).to.equal(1);
+    expect(timestring('1w', 'd', opts)).to.equal(2);
+    expect(timestring('1mth', 'w', opts)).to.equal(3);
+    expect(timestring('1y', 'mth', opts)).to.equal(4);
   });
 
   it('throws an error when an invalid unit is used in the timestring', function() {
-    var ts = new timestring();
-
-    expect(ts.parse.bind(ts, '1g')).to.throw(Error);
+    expect(function() { timestring('1g'); }).to.throw(Error);
   });
 
   it('can parse a messy time string', function() {
-    expect((new timestring()).parse('5   D a YS    4 h   2 0     mI  nS')).to.equal(447600);
-  });
-
-  it('should expose a method on String.prototype that will parse the string as a timestring', function(){
-    var str = '1min';
-
-    // no arguments passed
-    expect(str.parseTime()).to.equal(60);
-
-    // units argument passed
-    expect(str.parseTime('m')).to.equal(1);
-
-    // units + options argument passed
-    str = '5h';
-    expect(str.parseTime('d', { hoursPerDay: 5 })).to.equal(1);
+    expect(timestring('5   D a YS    4 h   2 0     mI  nS')).to.equal(447600);
   });
 });